fix unclear output when running git_pull_push.sh with `-x` execute command but no command to follow it: print a message there's no extra command given.
DOES NOT use this when we run other commands and the 'extra command' is just an *extra* feature: when it's not given (like usual), then keep silent about it, as before.

ALSO fix a bug/error in the script when running commands such as `gpp -0 -p`: the irst-character-extraction MAY nd up being empty and the [...] test statement could not handle that as we didn't wrap the empty value in string quotes.