Adding strip-all-but-original-remotes-and-clean.sh tool script: discards PERMANENTLY anything from the repo that's not already present in a REMOTE 'origin' or '*original*' remote: strip the repo down to the bare minimum and nuke all history, thus breaking up any previous relationships with git remotes/clones/forks elsewhere, except the 'original' source repo.
This effectively nukes ALL local and remote tags (labels) -- only a not-owned-by-me remote would be able to re-introduce those tags, plus it nukes all the commits that were introduced by other non-original registered remotes, all of whom will be deleted as well. (Yes, this script will nuke the registered remotes, so back up that list if you want to restore it afterwards -- though it won't be much use as the relation with those remotes will be broken up by then.)

Use this script to (permanently) reduce a repository down to minimal storage; may be used before squashing the entire repo into a single commit.